+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.3.3
 
 * Fix a crash on exit on iOS. See [#390]